# extras/openapi.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"Puddle Backend",
      "description":"\n    Backend API for Puddle, a dataset marketplace platform enabling\n    seamless interaction between dataset vendors and buyers, with AI agent\n    support for enhanced dataset processing and management. \n    Features include user authentication, dataset management, vector\n    similarity search, chat functionality, and comprehensive API\n    documentation.\n    ",
      "version":"1.0.0"
   },
   "paths":{
...
      "/api/v1/agents/":{
         "post":{
            "tags":[
               "agents"
            ],
            "summary":"Create a new AI agent",
            "description":"Create a new AI agent (Admin or Vendor only).",
            "operationId":"create_agent_api_v1_agents__post",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AgentCreate"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/AgentRead"
                        }
                     }
                  }
               },
               "404":{
                  "description":"Agent not found"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "get":{
            "tags":[
               "agents"
            ],
            "summary":"List AI agents",
            "description":"- Admin: list all agents.\n- Vendor: list only own agents.\n- Buyer: read-only list view of all agents (marketplace discovery).",
            "operationId":"list_agents_api_v1_agents__get",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"vendor_id",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"string",
                           "format":"uuid"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "description":"Filter by vendor ID",
                     "title":"Vendor Id"
                  },
                  "description":"Filter by vendor ID"
               },
               {
                  "name":"active",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "anyOf":[
                        {
                           "type":"boolean"
                        },
                        {
                           "type":"null"
                        }
                     ],
                     "description":"Filter by active status",
                     "title":"Active"
                  },
                  "description":"Filter by active status"
               },
               {
                  "name":"limit",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "maximum":1000,
                     "minimum":1,
                     "description":"Max number of agents",
                     "default":100,
                     "title":"Limit"
                  },
                  "description":"Max number of agents"
               },
               {
                  "name":"offset",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "minimum":0,
                     "description":"Number of agents to skip",
                     "default":0,
                     "title":"Offset"
                  },
                  "description":"Number of agents to skip"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/AgentRead"
                           },
                           "title":"Response List Agents Api V1 Agents  Get"
                        }
                     }
                  }
               },
               "404":{
                  "description":"Agent not found"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/v1/agents/{agent_id}":{
         "get":{
            "tags":[
               "agents"
            ],
            "summary":"Get a specific AI agent",
            "operationId":"get_agent_api_v1_agents__agent_id__get",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"agent_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "description":"The UUID of the agent",
                     "title":"Agent Id"
                  },
                  "description":"The UUID of the agent"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/AgentRead"
                        }
                     }
                  }
               },
               "404":{
                  "description":"Agent not found"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "put":{
            "tags":[
               "agents"
            ],
            "summary":"Update an AI agent",
            "description":"Update an AI agent's configuration (Admin or owning Vendor).",
            "operationId":"update_agent_api_v1_agents__agent_id__put",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"agent_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Agent Id"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "type":"object",
                        "additionalProperties":true,
                        "title":"Update Data"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/AgentRead"
                        }
                     }
                  }
               },
               "404":{
                  "description":"Agent not found"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "delete":{
            "tags":[
               "agents"
            ],
            "summary":"Delete an AI agent",
            "description":"Delete an AI agent (Admin or owning Vendor).",
            "operationId":"delete_agent_api_v1_agents__agent_id__delete",
            "security":[
               {
                  "OAuth2PasswordBearer":[
                     
                  ]
               }
            ],
            "parameters":[
               {
                  "name":"agent_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Agent Id"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "404":{
                  "description":"Agent not found"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components":{
      "schemas":{
         "AgentCreate":{
            "properties":{
               "vendor_id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Vendor Id"
               },
               "name":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Name"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "model_used":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Model Used",
                  "default":"gemini-embedding-001"
               },
               "config":{
                  "anyOf":[
                     {
                        "additionalProperties":true,
                        "type":"object"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Config"
               },
               "active":{
                  "anyOf":[
                     {
                        "type":"boolean"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Active",
                  "default":true
               }
            },
            "type":"object",
            "required":[
               "vendor_id"
            ],
            "title":"AgentCreate"
         },
         "AgentRead":{
            "properties":{
               "vendor_id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Vendor Id"
               },
               "name":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Name"
               },
               "description":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Description"
               },
               "model_used":{
                  "anyOf":[
                     {
                        "type":"string"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Model Used",
                  "default":"gemini-embedding-001"
               },
               "config":{
                  "anyOf":[
                     {
                        "additionalProperties":true,
                        "type":"object"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Config"
               },
               "active":{
                  "anyOf":[
                     {
                        "type":"boolean"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Active",
                  "default":true
               },
               "id":{
                  "type":"string",
                  "format":"uuid",
                  "title":"Id"
               },
               "created_at":{
                  "anyOf":[
                     {
                        "type":"string",
                        "format":"date-time"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Created At"
               },
               "updated_at":{
                  "anyOf":[
                     {
                        "type":"string",
                        "format":"date-time"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Updated At"
               }
            },
            "type":"object",
            "required":[
               "vendor_id",
               "id"
            ],
            "title":"AgentRead"
         },
...
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
...
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         },
...
      },
      "securitySchemes":{
         "OAuth2PasswordBearer":{
            "type":"oauth2",
            "flows":{
               "password":{
                  "scopes":{
                     
                  },
                  "tokenUrl":"/api/v1/auth/login"
               }
            }
         }
      }
   },
...
}